fixed some typos/syntax and renamed Handlers => ORM
